+++ HISTORY.md
@@ -10,6 +10,13 @@
 - Add PyTorch to ribs[all] deps ({pr}`692`)
 - Add `NSLCRanker` for Novelty Search with Local Competition ({pr}`690`)
 
+#### Documentation
+
+- Add script for running sphere over multiple trials ({pr}`696`)
+- Add Sphinx role for github badges ({pr}`697`)
+- Edits for tutorials and examples ({pr}`695`)
+- Update logging, outputs, and metrics in examples ({pr}`694`)
+
 ## 0.10.0
 
 ### Changelog